src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi[standard]>=0.135.1",
    "requests>=2.32.5",
    "google-cloud-bigquery>=3.11.0",
    "email-validator>=2.0.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
]
